Refactor exporter so that it can be used as a library. (#1)
